Fixed crash on push notification triggering overlapping navigation.
(APPS-854)

We want to only show the conversation associated with a received push
notification if the user tapped the notification. So we check for the
inactive state.

We also make sure that we don't have unnecessary navigation pushes by
checking if the conversation is already in the navigation stack.

Note that it's still technically possible to have overlapping navigation
pushes, e.g. if a user taps on a push notification, the app doesn't have
the conversation (A) yet so it begins to load it, meanwhile the user
taps on another conversation (B), conversation A finishes loading and
pushes the conversation view controller while the push for conversation
B is still in progress.