Fixed crash on sending and navigating from conversation and back.
(APPS-860)

When a user returns to the list of conversations, there may be a gap in
time between when pop navigation animation ends and the popped
conversation view controller is deallocated (with the duration seemingly
related to sending a message just before popping). So using a weak
property for the conversation view controller is an inaccurate technique
to determine whether or not there is an existing conversation view
controller in the navigation stack. We could continue using the
technique but with the addition of an explicit check to see if the
navigation controller's array of view controllers contains the
conversation view controller, but instead we simply get the next view
controller in the stack to make the check for obvious / self-contained.