Replaced Quick Look preview controller to work around iOS 8 bugs.
This fixes the background being transparent during the navigation
animation to the tapped image on iOS 8. (APPS-859)

This also fixes the navigation bar getting in an inconsistent state
after tapping an image to minimize the navigation and then tapping the
back button on iOS 8. (APPS-862) Note that this new view controller does
not allow minimizing the navigation bar.